plot the each anomaly sequences in a pdf file, respectively and plot the separate dim data in the same subplot

+++ src/smach_based_introspection_framework/offline_part/anomaly_classification_feature_selection/generate_dataset_plots.py
@@ -35,8 +35,9 @@ def run():
 
         output_dir = os.path.join(anomaly_classification_feature_selection_folder, "visualized dataset", path_postfix)
         csv_files = glob.glob(os.path.join(folder, '*', '*.csv'))
-        num_by_type = len(csv_files)
-        fig_whole_type, axarr = plt.subplots(nrows=num_by_type, ncols=1, figsize=(24, 6*num_by_type))        
+        csvs_by_type = len(csv_files)
+        fig_whole_type, axarr = plt.subplots(nrows=csvs_by_type, ncols=1, figsize=(24, 8*csvs_by_type))
+        flag = None
         for num, csv in enumerate(glob.glob(os.path.join(folder, '*', '*.csv'))):
     
             logger.info(csv)
@@ -51,41 +52,52 @@ def run():
                 os.makedirs(l2_output_dir)
 
             df = pd.read_csv(csv)
+            if flag is None:
+                flag = 'Enter'
+                ndims = len(df.columns)-1
+                fig_dims, axs = plt.subplots(nrows=ndims, ncols=1, figsize=(24, 6*ndims))
             for idx, col_name in enumerate(df.columns):
                 fig, ax = plt.subplots(nrows=1, ncols=1)
                 if idx == 0:
                     xs = df.iloc[:, 0]
-                    df.plot(ax=ax, x=df.keys()[0],y=df.keys()[1::])
-                    ax.legend(loc=2, fancybox=True, framealpha=0.3, ncol=1)
-                    if num_by_type <= 1:
-                        df.plot(ax=axarr, x=df.keys()[0], y=df.keys()[1::], title=exp_name , legend=True)
-                        axarr.legend(loc=2, fancybox=True, framealpha=0.3, ncol=1)                        
-                        axarr.axvline(anomaly_t, c='purple')
-                        ymin,ymax = axarr.get_ylim()
-                        axarr.text(anomaly_t, ymax-0.25*(ymax-ymin), '%s'%anomaly_type, color='purple', rotation=-90)                           
-                    else:
-                        df.plot(ax=axarr[num], x=df.keys()[0], y=df.keys()[1::], title=exp_name, legend=True)
-                        axarr[num].legend(loc=2, fancybox=True, framealpha=0.3, ncol=1)                        
-                        axarr[num].axvline(anomaly_t, c='purple')
-                        ymin,ymax = axarr[num].get_ylim()
-                        axarr[num].text(anomaly_t, ymax-0.25*(ymax-ymin), '%s'%anomaly_type, color='purple', rotation=-90)
-                    break
+                    
+                    axax = axarr[num] if csvs_by_type > 1 else axarr 
+                    df.plot(ax=axax, x=df.keys()[0], y=df.keys()[1::], title=exp_name, colormap='jet', legend=True)
+                    axax.legend(loc=2, framealpha=0.3, ncol=1)                        
+                    axax.axvline(anomaly_t, c='purple')
+                    ymin,ymax = axax.get_ylim()
+                    axax.text(anomaly_t, ymax-0.25*(ymax-ymin), '%s'%anomaly_type, color='purple', rotation=-90)
+                    continue
                 else:
                     ys = df.iloc[:, idx]
-                    ax.plot(xs, ys)                    
+                    ax.plot(xs, ys)
+                    if num == 0:
+                        xxs = xs
+                        axs[idx-1].plot(xxs, ys,label=col_name)
+                        axs[idx-1].legend(loc=2, fancybox=True, framealpha=0.3)
+                    elif num == csvs_by_type - 1:
+                        axs[idx-1].axvline(xxs.median(), c='purple')
+                        ymin,ymax = axs[idx-1].get_ylim()
+                        axs[idx-1].text(xxs.median(), ymax-0.25*(ymax-ymin), '%s'%anomaly_type, color='purple', rotation=-90)
+                    else:
+                        axs[idx-1].plot(xxs, ys)
+
+                output_f = os.path.join(l2_output_dir, "dim %s.png"%col_name.replace('/', 'divide'))
+                if os.path.exists(output_f):
+                    continue
                 xmin,xmax = ax.get_xlim()
                 ymin,ymax = ax.get_ylim()
-
                 ax.axvline(anomaly_t, c='purple')
                 ax.text(anomaly_t, ymax-0.25*(ymax-ymin), '%s'%anomaly_type, color='purple', rotation=-90)                    
                 ax.set_title("[%s] [%s] [%s]"%(anomaly_type, exp_name, col_name))
-                output_f = os.path.join(l2_output_dir, "dim %s.png"%col_name.replace('/', 'divide'))
                 fig.set_size_inches(16, 4)
                 with open(output_f, 'w') as f:
-                    fig.savefig(f)
+                    fig.savefig(f, dpi=300)
                 plt.close(fig)
         save_path = os.path.join(output_dir, anomaly_type + '.pdf')
         fig_whole_type.savefig(save_path, dpi=300)
+        dims_save_path = os.path.join(output_dir, anomaly_type + '_dims.pdf')
+        fig_dims.savefig(dims_save_path, dpi=300)
         
 
 if __name__ == '__main__':
